backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "research-agent-backend"
version = "0.1.0"
description = "Research agent backend powered by FastAPI, LangGraph, and LangChain."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "pydantic-settings>=2.4,<3.0",
  "google-genai>=1.0,<2.0",
  "langgraph>=0.2,<1.0",
  "langchain>=0.3,<1.0",
  "langchain-community>=0.3,<1.0",
  "langchain-text-splitters>=0.3,<1.0",
  "langsmith>=0.1,<1.0",
  "pinecone>=6.0,<7.0",
  "pypdf>=5.0,<6.0",
  "python-multipart>=0.0.9,<1.0",
  "orjson>=3.10,<4.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27,<1.0",
  "pytest>=8.3,<9.0"
]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
